Refacto : - renamed vtltape.h to vtlcart.h because it contains prototypes from both vtlcart.c and vtltape.c. Here, since vtltape.c is supposed to be relatively independant from the other binaries, it seems more fitting to have a vtlcart.h header that will be included in other files. - functions defined in vtltape.c that were called in other files were moved to ssc.c/ssc.h
