+++ usr/cmd/edit_tape.c
@@ -37,7 +37,7 @@
 #include "be_byteshift.h"
 #include "mhvtl_list.h"
 #include "vtl_common.h"
-#include "vtltape.h"
+#include "vtlcart.h"
 #include "vtllib.h"
 #include "logging.h"
