+++ usr/cmd/mktape.c
@@ -18,7 +18,7 @@
 #include "be_byteshift.h"
 #include "mhvtl_list.h"
 #include "vtl_common.h"
-#include "vtltape.h"
+#include "vtlcart.h"
 #include "vtllib.h"
 
 #if defined _LARGEFILE64_SOURCE
