+++ usr/cmd/tape_util.c
@@ -37,7 +37,7 @@
 #include "mhvtl_list.h"
 #include "vtl_common.h"
 #include "vtllib.h"
-#include "vtltape.h"
+#include "vtlcart.h"
 #include "q.h"
 #include "ssc.h"
 #include "ccan/crc32c/crc32c.h"
