+++ usr/pm/ait_pm.c
@@ -43,7 +43,7 @@
 #include "q.h"
 #include "ssc.h"
 #include "spc.h"
-#include "vtltape.h"
+#include "vtlcart.h"
 #include "q.h"
 #include "mode.h"
 #include "mhvtl_log.h"
