+++ usr/pm/t10000_pm.c
@@ -44,7 +44,7 @@
 #include "q.h"
 #include "ssc.h"
 #include "spc.h"
-#include "vtltape.h"
+#include "vtlcart.h"
 #include "mode.h"
 #include "mhvtl_log.h"
