+++ usr/vtlcart_v1_mtr.c
@@ -38,7 +38,7 @@
 
 #include "mhvtl_scsi.h"
 #include "mhvtl_list.h"
-#include "vtltape.h"
+#include "vtlcart.h"
 #include "be_byteshift.h"
 
 #define B_BOT 14 /* Beginning of Tape TAPE_FMT_VERSION 2 */
